Fix bug where hosts can't claim affinity properly.
This is a workaround for a design incompatibility between mesos and libcalico. Libcalico sometimes uses the hostname returned by socket.hostname, so if a host claims affinity for block under a different hostname, it can't claim it later. This fix prevents mesos from 'lying' about its hostname and using socket.gethostname itself, until libcalico can be updated to use a different hostname